fix(dsl): Change position type from int to float for ReactFlow compatibility
ReactFlow uses float coordinates, but generated Pydantic model required integers.
This caused validation errors when loading workflows with float positions.
Change position type to tuple[float, float] to match actual data.

+++ services/api/src/smith/schemas.py
@@ -15,7 +15,7 @@ class WorkflowNode(BaseModel):
     trigger: bool = False
     parameters: dict[str, Any] = Field(default_factory=dict)
     output: dict[str, Any] = Field(default_factory=dict)
-    position: Optional[tuple[int, int]] = None
+    position: Optional[tuple[float, float]] = None
 
 
 class WorkflowEdge(BaseModel):
